python-rabbitmq-volttron-client/requirements.txt
gevent-pika==0.2
gevent==1.1.2
PyYAML==3.12
requests[security]==2.21.0
